Refactor the big files.
